% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockplumber-package.R
\docType{package}
\name{mockplumber-package}
\alias{mockplumber}
\alias{mockplumber-package}
\title{mockplumber: Mock servers from 'API' documentation}
\description{
Create a 'plumber' file from an 'OpenAPI' document defining an 'API'. The 'plumber' file will have all the information to generate a mock server for that 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/mockplumber}
  \item \url{https://jonthegeek.github.io/mockplumber/}
  \item Report bugs at \url{https://github.com/jonthegeek/mockplumber/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
